Archived compare function
